+++ src/components/UploadCustomRecipe/useUploadCustom.js
@@ -10,7 +10,8 @@ const useUploadCustom = () => {
   const [ingredients, setIngredients] = useState("");
   const [method, setMethod] = useState("");
   const [duration, setDuration] = useState(0);
-  const [recipeObject, setRecipeObject] = useState({});
+  const [recipeObject, setRecipeObject] = useState("");
+  const [editRecipe, setEditRecipe] = useState(false);
 
   const recipeNameHandler = (e) => {
     setRecipeName(e.target.value);
@@ -26,11 +27,9 @@ const useUploadCustom = () => {
   };
 
   const selectedFileHandler = (e) => {
-     const file = e.target.files[0];
-      setSelectedFile(file)
+    const file = e.target.files[0];
+    setSelectedFile(file);
   };
-
-
 
   const ingredientsHandler = (e) => {
     setIngredients(e.target.value);
@@ -39,52 +38,44 @@ const useUploadCustom = () => {
   const methodHandler = (e) => {
     setMethod(e.target.value);
   };
+
+  const handleEditRecipe = () => {
+    setEditRecipe(true);
+  };
   const fileDataHandler = (e) => {
     e.preventDefault();
-    console.log("hello")
-    const id =  createId();
-  console.log( selectedFile);
+    const id = createId();
+
     recipesDB
-      .createRecipe({ id, recipeName, servings, ingredients, method, selectedFile, duration })
+      .createRecipe({
+        id,
+        recipeName,
+        servings,
+        ingredients,
+        method,
+        selectedFile,
+        duration,
+      })
       .catch((error) => {
         console.log("creation error", error);
       });
-
   };
 
-  const getRecipeObject =()=>{
-    return (recipesDB
-      .readRecipe("b056903e-988a-4e20-b00c-85752a8e8a05")
-    .then((result) => { 
-       console.log(result);
-      setRecipeObject({...result,photo: URL.createObjectURL(result.photo)})}))
-   
-  }
-  console.log(recipeObject)
-
-  // const RecipeObj =()=>{
-  //   {recipesDB
-  //     .readRecipe("b056903e-988a-4e20-b00c-85752a8e8a05")
-  //     .then((result) => {
-  //       const customRecipe = Object.keys(result);
-  //       console.log(Array.isArray(customRecipe),customRecipe);
-  //       Object.keys(result).map((item) => {
-  //         return (
-  //           <ul>
-  //             <li>{item.name}</li>
-  //             <li>{item.serves}</li>
-  //             <li>{item.photo}</li>
-  //             <li>{item.ingredients}</li>
-  //             <li>{item.method}</li>
-  //           </ul>
-  //         );
-  //       });
-  //     });
-  // } 
-    // 
+  const getRecipeObject = () => {
+    return recipesDB
+      .readRecipe("5925e4ba-896a-4034-a959-ef314b854889")
+      .then((result) => {
+        console.log(result);
+        setRecipeObject({
+          ...result,
+          photo: URL.createObjectURL(result.photo),
+        });
+      });
+  };
+  console.log(recipeObject);
 
   return {
-     getRecipeObject,
+    getRecipeObject,
     recipeObject,
     method,
     methodHandler,
@@ -98,6 +89,8 @@ const useUploadCustom = () => {
     recipeName,
     recipeNameHandler,
     fileDataHandler,
+    handleEditRecipe,
+    editRecipe,
   };
 };
